Struct prost_reflect::EnumValueDescriptor

pub struct EnumValueDescriptor { /* private fields */ }

A value in a protobuf enum type.

Implementations

impl EnumValueDescriptor

pub fn parent_pool(&self) -> &DescriptorPool

Gets a reference to the DescriptorPool this enum value is defined in.

pub fn parent_file(&self) -> FileDescriptor

Gets the FileDescriptor this enum value is defined in.

pub fn parent_enum(&self) -> &EnumDescriptor

Gets a reference to the EnumDescriptor this enum value is defined in.

pub fn name(&self) -> &str

Gets the short name of the enum value, e.g. MY_VALUE.

pub fn full_name(&self) -> &str

Gets the full name of the enum value, e.g. my.package.MY_VALUE.

pub fn path(&self) -> &[i32]

Gets the path where this enum value is defined within the FileDescriptorProto, e.g. [5, 0, 2, 0].

See path for more details on the structure of the path.

pub fn enum_value_descriptor_proto(&self) -> &EnumValueDescriptorProto

Gets a reference to the raw EnumValueDescriptorProto wrapped by this EnumValueDescriptor.

pub fn options(&self) -> DynamicMessage

Decodes the options defined for this EnumValueDescriptor, including any extension options.

pub fn number(&self) -> i32

Gets the number representing this enum value.