Struct prost_reflect::FieldDescriptor

pub struct FieldDescriptor { /* private fields */ }

A protobuf message definition.

Implementations

impl FieldDescriptor

pub fn parent_pool(&self) -> &DescriptorPool

Gets a reference to the DescriptorPool this field is defined in.

pub fn parent_file(&self) -> FileDescriptor

Gets the FileDescriptor this field is defined in.

pub fn parent_message(&self) -> &MessageDescriptor

Gets a reference to the MessageDescriptor this field is defined in.

pub fn name(&self) -> &str

Gets the short name of the message type, e.g. my_field.

pub fn full_name(&self) -> &str

Gets the full name of the message field, e.g. my.package.MyMessage.my_field.

pub fn path(&self) -> &[i32]

Gets the path where this message field is defined within the FileDescriptorProto, e.g. [4, 0, 2, 0].

See path for more details on the structure of the path.

pub fn field_descriptor_proto(&self) -> &FieldDescriptorProto

Gets a reference to the raw FieldDescriptorProto wrapped by this FieldDescriptor.

pub fn options(&self) -> DynamicMessage

Decodes the options defined for this FieldDescriptor, including any extension options.

pub fn number(&self) -> u32

Gets the unique number for this message field.

pub fn json_name(&self) -> &str

Gets the name used for JSON serialization.

This is usually the camel-cased form of the field name, unless another value is set in the proto file.

pub fn is_group(&self) -> bool

Whether this field is encoded using the proto2 group encoding.

pub fn is_list(&self) -> bool

Whether this field is a list type.

Equivalent to checking that the cardinality is Repeated and that is_map returns false.

pub fn is_map(&self) -> bool

Whether this field is a map type.

Equivalent to checking that the cardinality is Repeated and that the field type is a message where is_map_entry returns true.

pub fn is_packed(&self) -> bool

Whether this field is a list encoded using packed encoding.

pub fn cardinality(&self) -> Cardinality

The cardinality of this field.

pub fn supports_presence(&self) -> bool

Whether this field supports distinguishing between an unpopulated field and the default value.

For proto2 messages this returns true for all non-repeated fields. For proto3 this returns true for message fields, and fields contained in a oneof.

pub fn kind(&self) -> Kind

Gets the Kind of this field.

pub fn containing_oneof(&self) -> Option<OneofDescriptor>

Gets a OneofDescriptor representing the oneof containing this field, or None if this field is not contained in a oneof.